+++ src/app/_components/Header.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import { UserButton, useAuth, OrganizationSwitcher } from "@clerk/nextjs";
+import { UserButton, useUser, OrganizationSwitcher } from "@clerk/nextjs";
 import Image from "next/image";
 import Link from "next/link";
 import { useRouter, usePathname } from "next/navigation";
 import { env } from "@/env.mjs";
-import { useEffect, useState } from "react";
+import Loading from "./Loading";
 
 interface NavbarProps {
   title: string;
@@ -14,15 +14,9 @@ interface NavbarProps {
 export const dynamic = "force-dynamic";
 
 const Navbar = ({ title }: NavbarProps) => {
-  const { isSignedIn } = useAuth();
+  const { isLoaded, isSignedIn } = useUser();
   const router = useRouter();
   const pathname = usePathname();
-
-  const [isReady, setIsReady] = useState(false);
-
-  useEffect(() => {
-    setIsReady(true);
-  }, []);
 
   const navigationMenu = () => {
     if (pathname !== "/dashboard" && isSignedIn) {
@@ -46,6 +40,7 @@ const Navbar = ({ title }: NavbarProps) => {
           createOrganizationMode="modal"
           afterCreateOrganizationUrl="/dashboard"
           afterLeaveOrganizationUrl="/dashboard"
+          afterSwitchOrganizationUrl="/dashboard"
           appearance={{
             elements: {
               organizationSwitcherTrigger: "text-white",
@@ -80,12 +75,9 @@ const Navbar = ({ title }: NavbarProps) => {
           </span>
         </Link>
       </div>
-      {isReady && (
-        <>
-          {navigationMenu()}
-          <UserButton afterSignOutUrl="/" userProfileMode="modal" />
-        </>
-      )}
+
+      {!isLoaded ? <Loading /> : navigationMenu()}
+      <UserButton afterSignOutUrl="/" userProfileMode="modal" />
     </nav>
   );
 };
